🛠️ :: (#276) 연매출 이름 take로 통일, 실수형으로 변경

+++ src/main/java/team/returm/jobis/domain/company/presentation/dto/request/RegisterCompanyRequest.java
@@ -79,7 +79,7 @@ public class RegisterCompanyRequest {
     private int workerNumber;
 
     @NotNull
-    private int take;
+    private double take;
 
     @NotBlank
     private String companyProfileUrl;
